#스텟 클래스 수정

+++ Assets/Script/GameEvent.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public static class GameEvents
 {
     public static Action<int> OnGoldChanged;
-    public static Action<int> OnPlayerHPChanged;
+    public static Action<float> OnPlayerHPChanged;
     public static Action<int> OnWaveChanged;
     public static Action OnWaveStart;
     public static Action OnPlayerDeath;
